// app/javascript/api/types/rails/listings/RailsListingPricingTableUnit.d.ts
export type RailsListingPricingTableUnit = {
  unitType: string
  unitMaxAMI: number
  totalUnits: number
  minSquareFt: number
  minPriceWithParking: number
  minPriceWithoutParking: number
  minPercentIncome: number
  minOccupancy: number
  minMonthlyRent: number
  minHoaDuesWithParking: number
  minHoaDuesWithoutParking: number
  maxSquareFt: number
  maxRentalMinIncome: number
  maxPriceWithParking: number
  maxPriceWithoutParking: number
  maxOccupancy: number
  maxMonthlyRent: number
  maxHoaDuesWithParking: number
  maxHoaDuesWithoutParking: number
  listingID: string
  availability: number
  absoluteMinIncome: number
  absoluteMaxIncome: number
  subsetValue: number
  minRentalMinIncome: number
  maxPercentIncome: number
}